SelectBoost.beta: Stability-Selection via Correlated Resampling for Beta-Regression Models

Adds variable-selection functions for Beta regression models (both mean and phi submodels) so they can be used within the 'SelectBoost' algorithm. Includes stepwise 'AIC'/'BIC'/'AICc' on betareg() fits, 'GAMLSS'-based LASSO/Elastic-Net, a pure 'glmnet' 'IRLS'-based selector with an optional standardization speedup, and C++ helpers for 'IRLS' working steps and precision updates. Also provides a fastboost_interval() variant for interval responses, comparison helpers, and a flexible simulator simulation_DATA.beta() for interval-valued data. The 'SelectBoost4Beta' framework was presented by Bertrand and 'Maumy' (2023) at Joint Statistical Meetings 2023 (Toronto, Canada) and 'BioC2023' (Boston, USA), highlighting correlated resampling for Beta-regression variable selection.

References

Bertrand, F. and Maumy, M. (2023). Improving variable selection in Beta regression models using correlated resampling. Presented at Joint Statistical Meetings 2023, Toronto, Canada.

Bertrand, F. and Maumy, M. (2023). SelectBoost4Beta: Improving variable selection in Beta regression models. Presented at BioC2023, Boston, USA.

See also

Author

This package was written by Frédéric Bertrand. Maintainer: Frédéric Bertrand frederic.bertrand@lecnam.net

Examples

set.seed(1)
n <- 150; p <- 6
X <- matrix(rnorm(n*p), n, p); colnames(X) <- paste0("x",1:p)
eta <- 0.4 + X[,1] - 0.7*X[,3]
mu  <- plogis(eta)
Y   <- rbeta(n, mu*25, (1-mu)*25)

betareg_step_aic(X, Y)    # should return (Intercept) + x1,x3 nonzero often
#>     (Intercept)              x1              x2              x3              x4 
#>       0.3840811       1.0104247       0.0000000      -0.6555926       0.0000000 
#>              x5              x6 phi|(Intercept) 
#>       0.0000000       0.0000000       3.2252600 
betareg_step_bic(X, Y)
#>     (Intercept)              x1              x2              x3              x4 
#>       0.3840811       1.0104247       0.0000000      -0.6555926       0.0000000 
#>              x5              x6 phi|(Intercept) 
#>       0.0000000       0.0000000       3.2252600 
betareg_step_aicc(X, Y)
#>     (Intercept)              x1              x2              x3              x4 
#>       0.3840811       1.0104247       0.0000000      -0.6555926       0.0000000 
#>              x5              x6 phi|(Intercept) 
#>       0.0000000       0.0000000       3.2252600