Core helpers for SelectBoost-style beta regression

These helpers expose the individual stages of the SelectBoost workflow so that beta-regression selectors can be combined with correlation-aware resampling directly from SelectBoost.beta. They normalise the design matrix, derive correlation structures, form groups of correlated predictors, generate Gaussian surrogates that mimic the observed dependency structure, and apply a user-provided selector on each resampled design.

Usage

sb_normalize(X, eps = 1e-08)

sb_compute_corr(X, corrfunc = "cor")

sb_group_variables(corr_mat, c0)

Arguments

X

Numeric matrix of predictors.

eps

Small positive constant used when normalising columns.

corrfunc

Function or character string used to compute pairwise associations. Defaults to "cor".

corr_mat

Numeric matrix of associations.

c0

Threshold applied to the absolute correlations.

Value

sb_normalize() returns a centred, \(\ell_2\)-scaled copy of X.

sb_compute_corr() returns the association matrix.

sb_group_variables() returns a list of integer vectors, one per variable, describing the correlated group it belongs to.

Examples

sb_normalize(matrix(rnorm(20), 5))
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.5783289  0.33811804 -0.69163269  0.1982533
#> [2,]  0.5157848 -0.02726555 -0.29829359  0.7439537
#> [3,] -0.3469389 -0.31850801  0.51782060 -0.5281400
#> [4,] -0.3732982  0.62970959  0.07313399 -0.0611383
#> [5,] -0.3738765 -0.62205407  0.39897168 -0.3529287